pyproject.toml
[tool.poetry]
name = "cupapp2"
version = "0.1.0"
description = "CupApp is an django-based web application written for manage football cups organized among friends."
authors = ["sequencer <sequencer@o2.pl>"]
license = "MIT Licence"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
PyYAML = "^6.0"
pytz = "^2021.3"

[tool.poetry.dev-dependencies]
parameterized = "^0.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
